change order number type

+++ internal/adapters/storage/orders.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 )
 
-func (s *Storage) FindUserIDByOrderNumber(ctx context.Context, number int) (int, error) {
+func (s *Storage) FindUserIDByOrderNumber(ctx context.Context, number int64) (int, error) {
 	var userID int
 
 	q := `SELECT user_id FROM doc_order WHERE number = $1 LIMIT 1`
@@ -23,7 +23,7 @@ func (s *Storage) FindUserIDByOrderNumber(ctx context.Context, number int) (int,
 	return userID, nil
 }
 
-func (s *Storage) CreateOrder(ctx context.Context, number int, userID int) error {
+func (s *Storage) CreateOrder(ctx context.Context, number int64, userID int) error {
 	q := `INSERT INTO doc_order (number, user_id) VALUES ($1, $2)`
 
 	_, err := s.db.ExecContext(ctx, q, number, userID)
