+++ migrations/20231014171759_doc_order_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABlE doc_order
 (
     id         SERIAL PRIMARY KEY,
-    number     INT UNIQUE,
+    number     BIGINT UNIQUE,
     user_id    INT REFERENCES ref_user (id) NOT NULL,
     status     VARCHAR(100)                 NOT NULL DEFAULT 'NEW',
     changed_at TIMESTAMP(0) WITH TIME ZONE           DEFAULT now(),
